updated data processor and metrics
