+++ pytorch_transformers/data/processors/__init__.py
@@ -1,0 +1,2 @@
+from .utils import InputExample, InputFeatures, DataProcessor
+from .glue import output_modes, processors, convert_examples_to_glue_features
